Refactor blog generator
The blog generator was split into three separate generators: feed, index and tags
Besides code reogranization, some obsolete code and useless functionality was deleted.

Fixed #81
